change sampling mask to save as a single, since reading a mode 0 is somehow 5x slower. I don't have time to figure that out right now. also added a flag to PCA so that the regular bandpass is used unless a specific flag is added to the param file.

+++ ctf/BH_ctf_Updatefft.m
@@ -352,7 +352,7 @@ system('mkdir -p aliStacks');
   sizeCropped(3) = d3; 
   
  STACK = zeros(sizeCropped,'single');
- samplingMaskStack = zeros(sizeCropped,'uint8');
+ samplingMaskStack = zeros(sizeCropped,'single');
  
   for i = 1:d3
 
@@ -547,7 +547,7 @@ system('mkdir -p aliStacks');
                                              'GPU','single')));
                                            
       iSamplingMask(isnan(iSamplingMask(:))) = 0;
-      samplingMaskStack(:,:,i) = uint8(gather(real(iSamplingMask)));
+      samplingMaskStack(:,:,i) = (gather(real(iSamplingMask)));
       iSamplingMask = [];
 %      
 %   end
